refactor: encapsulate login and register component to auth component
